support for large files

+++ compressKernel.h
@@ -11,8 +11,8 @@ __global__ void cu_histgram(unsigned int *d_PartialHistograms,
 __global__ void mergeHistogram(unsigned int *d_Histogram,
                                unsigned int *d_PartialHistograms);
 
-__global__ void encode(uint fileSize, uint *dfileContent, uint *dbitOffsets,
-                       uint *d_boundary_index, uint *d_compressedFile,
+__global__ void encode(uint fileSize, uint *dfileContent, unsigned long long int *dbitOffsets,
+                       unsigned long long int *d_boundary_index, uint *d_compressedFile,
                        uint *d_dictionary_code,
                        unsigned char *d_dictionary_codelens, uint *counter,
                        uint numTasks);
